feat: kleros module has monitoring
kleros reality module now has two steps,
along with an optional monitoring setup.

some monitoring logic was reused from
current reality module.
notably, emails was redone because it
didn't seem obvious to use a multi-select
component.

mui-switch was hard to work with and i
could not figure out how to style it according
to zodiac style.

the env infura key is used to fetch ens
information, instead of relying on safe provider,
which could be on a non-mainnet chain.

this commit has not been tested because
i could not figure out how to boot the backend.
as a pragmatic workaround, im making a PR to
let CI/CD deploy it for me, and I'll test it there

non-obvious changes:
- installed lodash.debounce
- created an alternative checkmark icon
without hardcoded black fill